requirements.txt
aiofiles==23.2.1
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.13.1
Flask==3.0.0
frozenlist==1.4.0
fsspec==2023.10.0
gigachain==0.0.343
gigachain-cli==0.0.19
gigachain-core==0.0.7
gigachat==0.1.8
gigaserve==0.0.27
gitdb==4.0.11
GitPython==3.1.40
grpcio==1.59.3
grpcio-tools==1.59.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
httpx-sse==0.3.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langsmith==0.0.67
llvmlite==0.41.1
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
orjson==3.9.10
outcome==1.3.0.post0
packaging==23.2
pocketsphinx==5.0.2
protobuf==4.25.1
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
PySocks==1.7.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rich==13.7.0
selenium==4.15.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.23
srt==3.5.3
sse-starlette==1.8.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tomlkit==0.12.3
torch==2.1.1
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.23.2
vosk==0.3.44
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.3
